Refactor tests to mock size validator
